+++ ui/src/components/merge/MergePane.js
@@ -1,5 +1,6 @@
 import {faArrowRightLong} from "@fortawesome/free-solid-svg-icons";
 import MergeButton from "@/components/merge/MergeButton";
+import SearchableSelect from "@/components/SearchableSelect";
 import {useContext, useEffect, useState} from "react";
 import Modal from "@/components/Modal";
 import AppContext from "@/components/AppContext";
@@ -77,7 +78,9 @@ export default function MergePane({leftContext, rightContext, diff_type, merging
               <label className="select-all" style={{
                 display: selectAllVisible ? "inline-block" : "none"
               }}>
-                <input className="form-check-input" type="checkbox" checked={mergingContext.selectAll} onChange={() => {
+                <input className="form-check-input" type="checkbox"
+                       ref={(el) => { if (el) { el.indeterminate = mergingContext.selectionCount > 0 && !mergingContext.selectAll; } }}
+                       checked={mergingContext.selectAll} onChange={() => {
                   mergingContext.setAndApplySelectAll(!mergingContext.selectAll);
                 }}/>
                 select all
@@ -87,11 +90,11 @@ export default function MergePane({leftContext, rightContext, diff_type, merging
             {rightContext && !rightContext.authorizedForMerge && <span>You are not authorized to merge into target project</span>}
             {!mergeDisabled && (diff_type === DiffTypes.DOCUMENTS_DIFF || diff_type === DiffTypes.WORK_ITEMS_DIFF) && <label className="link-role-direction-label">
               <span className="link-role-direction-text">Link role direction for created WorkItems:</span>
-              <select className="form-select link-role-direction-select" value={linkRoleDirection}
-                      onChange={(e) => setLinkRoleDirection(e.target.value)} title="Link role direction for created WorkItems">
+              <SearchableSelect className="form-select link-role-direction-select" value={linkRoleDirection}
+                                onChange={(e) => setLinkRoleDirection(e.target.value)} searchable={false}>
                 <option value={LINK_ROLE_DIRECTION_DIRECT}>Direct</option>
                 <option value={LINK_ROLE_DIRECTION_REVERSE}>Reverse</option>
-              </select>
+              </SearchableSelect>
             </label>}
             <MergeButton fontAwesomeIcon={faArrowRightLong} clickHandler={() => confirmMerge(LEFT_TO_RIGHT)}
                          style={{justifyContent: "right"}} disabled={mergeDisabled || mergingContext.selectionCount === 0} />